Added direction for how the system should source control managers.
The system will only support Git for now, but that doesn't imply that it'll only support Git. Overall,  once we've implemented a parsing logic for a template file that we'll create carefully and slowly based on the example provided in the comments, we'll be able to perform virtually any SCM command from this tool.

The utility here comes from having a DSL that knows what we're trying achieve and expresses the intent into a CI domain related to the gaming industry. This serves as an example of what can be done. It could be done in a very different way most probably.